feat : Core Data 모델 생성, DataManager클래스 생성 - AppDelegate 메소드 core data stack, saveContext DataManager클래스로 이동
