+++ CloudNotes/CloudNotes/SceneDelegate.swift
@@ -33,7 +33,7 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
     }
 
     func sceneDidEnterBackground(_ scene: UIScene) {
-        (UIApplication.shared.delegate as? AppDelegate)?.saveContext()
+        DataManager.shared.saveContext()
     }
 
 }
